Рефакторинг кода масштабирования
